requirements.txt
python-dotenv==1.0.0
discord.py==2.3.2
openai==1.13.3
google-api-python-client==2.108.0
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
moviepy==1.0.3
Pillow==10.2.0
requests==2.31.0
python-slugify==8.0.1
